# Change JsonException.Path to be a json-based path