--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.cs
@@ -106,7 +106,7 @@ namespace System.Text.Json.Serialization
             catch (JsonReaderException e)
             {
                 // Re-throw with Path information.
-                ThrowHelper.ReThrowWithPath(e, state.PropertyPath);
+                ThrowHelper.ReThrowWithPath(e, state.JsonPath);
             }
 
             return;