--- a/src/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 
@@ -10,7 +11,9 @@ namespace System.Text.Json.Serialization
     [DebuggerDisplay("Current: ClassType.{Current.JsonClassInfo.ClassType}, {Current.JsonClassInfo.Type.Name}")]
     internal struct ReadStack
     {
-        // A fields is used instead of a property to avoid value semantics.
+        private static readonly char[] SpecialCharacters = { '.', ' ', '\'', '/', '"', '[', ']', '(', ')', '\t', '\n', '\r', '\f', '\b', '\\', '\u0085', '\u2028', '\u2029' };
+
+        // A field is used instead of a property to avoid value semantics.
         public ReadStackFrame Current;
 
         private List<ReadStackFrame> _previous;
@@ -48,47 +51,99 @@ namespace System.Text.Json.Serialization
 
         public bool IsLastFrame => _index == 0;
 
-        // Return a property path in the form of: [FullNameOfType].FirstProperty.SecondProperty.LastProperty
-        public string PropertyPath
+        // Return a JSONPath using simple dot-notation when possible. When special characters are present, bracket-notation is used:
+        // $.x.y[0].z
+        // $.['PropertyName.With.Special.Chars']
+        public string JsonPath
         {
             get
             {
-                StringBuilder path;
+                StringBuilder sb = new StringBuilder("$");
 
-                if (_previous == null || _index == 0)
+                for (int i = 0; i < _index; i++)
                 {
-                    // No path if we've walked beyond the end of our JSON document
-                    if (Current.JsonClassInfo == null)
+                    AppendStackFrame(sb, _previous[i]);
+                }
+
+                AppendStackFrame(sb, Current);
+                return sb.ToString();
+            }
+        }
+
+        private void AppendStackFrame(StringBuilder sb, in ReadStackFrame frame)
+        {
+            // Append the property name.
+            string propertyName = GetPropertyName(frame);
+            AppendPropertyName(sb, propertyName);
+
+            if (frame.JsonClassInfo != null)
+            {
+                if (frame.IsProcessingDictionary)
+                {
+                    // For dictionaries add the key.
+                    AppendPropertyName(sb, frame.KeyName);
+                }
+                else if (frame.IsProcessingEnumerable)
+                {
+                    // For enumerables add the index.
+                    IList list = frame.TempEnumerableValues;
+                    if (list == null && frame.ReturnValue != null)
                     {
-                        return "<none>";
+                        list = (IList)frame.JsonPropertyInfo?.GetValueAsObject(frame.ReturnValue);
                     }
 
-                    path = new StringBuilder($"[{Current.JsonClassInfo.Type.FullName}]");
+                    if (list != null)
+                    {
+                        sb.Append(@"[");
+                        sb.Append(list.Count);
+                        sb.Append(@"]");
+                    }
+                }
+            }
+        }
+
+        private void AppendPropertyName(StringBuilder sb, string propertyName)
+        {
+            if (propertyName != null)
+            {
+                JsonEncodedText encodedPropertyName = JsonEncodedText.Encode(propertyName);
+
+                if (propertyName.IndexOfAny(SpecialCharacters) != -1)
+                {
+                    sb.Append(@"['");
+                    sb.Append(encodedPropertyName);
+                    sb.Append(@"']");
                 }
                 else
                 {
-                    path = new StringBuilder($"[{_previous[0].JsonClassInfo.Type.FullName}]");
-
-                    for (int i = 0; i < _index; i++)
-                    {
-                        path.Append(GetPropertyName(_previous[i]));
-                    }
+                    sb.Append('.');
+                    sb.Append(encodedPropertyName);
                 }
-
-                path.Append(GetPropertyName(Current));
-
-                return path.ToString();
             }
         }
 
         private string GetPropertyName(in ReadStackFrame frame)
         {
-            if (frame.JsonPropertyInfo?.PropertyInfo != null && frame.JsonClassInfo.ClassType == ClassType.Object)
+            // Attempt to get the JSON property name from the frame.
+            byte[] utf8PropertyName = frame.JsonPropertyName;
+            if (utf8PropertyName == null)
             {
-                return $".{frame.JsonPropertyInfo.PropertyInfo.Name}";
+                // Attempt to get the JSON property name from the JsonPropertyInfo.
+                utf8PropertyName = frame.JsonPropertyInfo?.JsonPropertyName;
             }
 
-            return string.Empty;
+            string propertyName;
+            if (utf8PropertyName != null)
+            {
+                // Attempt to get the JSON property name from the dictionary key.
+                propertyName = JsonHelpers.Utf8GetString(utf8PropertyName);
+            }
+            else
+            {
+                propertyName = null;
+            }
+
+            return propertyName;
         }
     }
 }