--- a/src/System.Text.Json/tests/Serialization/Value.ReadTests.cs
+++ b/src/System.Text.Json/tests/Serialization/Value.ReadTests.cs
@@ -379,7 +379,7 @@ namespace System.Text.Json.Serialization.Tests
             {
                 exceptionThrown = true;
 
-                // Exception should contain property path.
+                // Exception should contain path.
                 Assert.True(exception.ToString().Contains("Path: $.Children[0].MyProperty"));
             }