Fixes https://github.com/dotnet/corefx/issues/37768

Currently we escape the property names in the Path.

We need to be clear on what is reported on invalid array element. Do we return the invalid array element, or the "last known good" element. For example does `[1,X` report a Path of `$[0]` (last known good) or `$[1]` (invalid element even though it's not a valid indexer). Currently the code returns `$[1]` for this case. _UPDATE: this behavior has been approved._

The path is not followed into calls into `JsonDocument` which occurs during deserialization of type(object) and for the overflow\[DataExtensionProperty]. The current plan is to create another issue to address this by first adding support within `JsonDocument` to support Path (which would work when not called by the serializer), and then have the serializer catch the appropriate `JsonReaderException` thrown by `JsonDocument` and prepend the exception's Path with the Path that existed before the call to `JsonDocument` and then re-throw as `JsonException`. _UPDATE: see issue https://github.com/dotnet/corefx/issues/38074_