@steveharter 2019-05-30
@ahsonkhan I previously measured benchmarks -- with case-sensitive the results didn't change (within margin of error). I re-ran locally with threshold of 1% and results are consistent with that:

```
better: 9, geomean: 1.250
worse: 6, geomean: 1.183
total diff: 15

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality  |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToString |      1.43 |           612.14 |           876.19 |           |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.19 |         49452.33 |         58921.34 | bimodal   |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.18 |        698441.79 |        826713.44 | bimodal   |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.16 |        745253.35 |        868183.78 |           |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.11 |        725348.65 |        802737.53 | multimodal|
| System.Text.Json.Serialization.Tests.WriteJson<IndexViewModel>.SerializeToStream |      1.05 |         34075.94 |         35865.59 |           |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality  |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromStr |      1.45 |          1192.24 |           821.82 | bimodal   |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromStr |      1.36 |          1571.40 |          1151.44 | several?  |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromStream    |      1.31 |          3423.60 |          2623.12 | several?  |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromUtf |      1.29 |          1018.41 |           791.12 | bimodal   |
| System.Text.Json.Serialization.Tests.WriteJson<IndexViewModel>.SerializeToString |      1.20 |         45131.60 |         37657.88 | multimodal|
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromString    |      1.18 |          2649.39 |          2249.76 |           |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.18 |        609012.88 |        517570.93 | bimodal   |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToUtf8By |      1.17 |           651.04 |           554.58 | bimodal   |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromUtf8Bytes |      1.15 |          2239.36 |          1945.41 | bimodal   |
```