@JamesNK 2019-05-30
https://github.com/dotnet/corefx/pull/37938#discussion_r289150816

https://github.com/dotnet/corefx/pull/37938#discussion_r289151573

My feedback??